Updated install_es_indexes management command
We should be referencing aliases from our code, not indices. This change updates the management command to create an index with a unique name and assign it an alias. Documentation has been updated accordingly. Additionally, the code creating the alias/index has been moved to a class that is now shared with the tests to avoid code duplication and ensure tests are run under similar conditions (e.g. alias pointing to index) to production.