Use single shard for catalog index.
The catalog index is small in terms of number of documents and data
size. There is no reason to prematurely shard the index. In addition,
sharding seems to be affecting our ability to properly score search
results with such a small data set causing inconsistent results when
performing queries which match many documents in the index.

ENT-444