Add an introduction to docs
This is the first in a series of commits that aims to update this documentation in an effort to make it more useful to other developers and the Open edX community at large. The introduction provides some background information on the discovery service that was missing before. It also explains key concepts necessary to understand the service. This change also brings these docs into line with those written for other services by switching to using the edX theme.

LEARNER-1399