Return distinct list of available courses.
The query that fetches available courses performs a join across the
course table, the course_runs table, and the seats table and returns
a separate copy of the same course for each available run/seat
combination by default.

We have to use distinct() to make sure only a single copy of each
available course is returned.