Push products to ecom in new publisher
Consolidate the 'publish to ecommerce' logic between old and new
publishers into a common code flow. Then use it in both places.

Also fixes the Program.entitlements utility function to actually
return the entitlements in a program, which is used when calculating
program price, in conjunction with the seat prices.

DISCO-984