Change CompressedCacheMixin strategy
Changes the compression/caching strategy to separate the response into
a tuple, and compress the rendered content, before reassembling the
response when it comes out of the cache. This is done because rendered
content can't actually be set.

The CompressedCacheResponseMixin is removed from all endpoints where
it is used so that we can remove all keys from the cache, then add the
mixin back in, hopefully preventing any issues with different data
formats in the cache.