DISCO-1082: Ensure audit entitlement is created
When an official course has no entitlement, it might be an audit course
and so we will set a default of creating a draft audit entitlement. This
happens as part of ensure_draft_world and now we ensure the update code
does not error if no entitlement data is passed in.