Add key_for_reruns field to Courses
This field will specify the course key used when creating new
reruns under the course. This allows our course teams to "soft
rename" course keys.

Their current approach for this is to create a run under a course
with the desired key. Then move the run to the original course.
Hopefully this will be easier for everyone.

DISCO-1358