Use Mode to determine ecom attributes
For seats, we used to set is_id_verified and certificate_type ecom
attributes by string comparision on seat type. But now we have Mode,
which holds those flags.

So now we send those along now based on Mode. The only behavior
difference is that we won't send any information for seats that
aren't supposed to be in the that run's CourseRunType.