Upgrade empty course & run types in RCM
After loading seats from ecommerce, try filling in any empty course
or course run types. This should only happen if we import runs from
Studio, in practice.

If a run has seats but the course or course run does not have a type
and we fail to find a matching type to use, we will fail the ecom
loader (but will let it finish first).

DISCO-1427