Stop unpublishing all runs in a course
We accidentally were unpublishing all inactive runs in a course,
even if that left us with no published runs. But we are supposed
to always leave at least one published run so that a course is
still marketable.

This fixes that logic and adds new code to the
unpublish_inactive_runs management command to re-publish old runs
that we accidentally unpublished.