Update dedupe script to use the `everything` manager
Several models in this repo inherit the DraftModelMixin which overrides
the `objects` attribute to only return non-draft objects, but also
provides an `everything` model manager which returns all objects.  This
change to the deduplicate_course_metadata_history management command
makes use of the `everything` model manager so that draft instances of a
model are not skipped while deduplicating history.