Fixing RCM
RCM has been failing because it has not had access to all of the
courses because the service user was being treated as an
AnonymousUser instead of a staff user. By passing in the username,
it will be treated as staff and get all courses.

We are also fixing our backoff code that started failing when we
updated which client we use.