Optimize the performance of the CourseQuerySet
The CourseQuerySet.available() function is used frequently, and
does a distinct to only return one copy of each Course. However,
because of the query construction, it needed to do a SELECT DISTINCT
over all fields of the Course (and associated Course Runs).

The new version just fetches distinct Course ids, and then gets all
course metadata for those courses. This improves query times by about
a factor of 3 on the read-replica.