We need to delete old entitlements if the Course Type changed
We saw a scenario where a draft Course started as Verified and Audit
and had a draft verified Entitlement as a result. The draft Course
was then switched to Audit Only, but we failed to remove the
Verified entitlement causing it to be created in Ecommerce. This
will prevent that scenario and ensure proper cleanup.

Note: This does not effect the behavior of published Courses and
Course runs as they are controlled by different logic to not allow
changing of entitlements