Fix LMS read timeout on running migrations
When running the install_es_indexes command, the discovery app connects
with the LMS. The default read timeout of 1s is sometimes too low, and
the command crashes. Here, we make use of the `timeout=` keyword
argument of the  OAuthAPIClient constructor to manually set the timeout.
We set it in the django settings to the default value of the class
constructor, such that existing platforms are not affected, but this
setting is customizable across environments.

See:
https://discuss.overhang.io/t/error-during-new-setup-no-module-named-ecommerce-settings-tutor/681/10
https://github.com/edx/edx-rest-api-client/pull/62#issuecomment-646526313