Add uniqueness constraints on Organization.uuid and .key
Throughout the Platform, we make assumptions around the
uniqueness of they Organization.key field (known as
Organization.short_name in edx-platform). As part of the
phase in of db-backed-organizations for all edx-platform
instances, it would be best to enforce this uniqueness at
the database level.

Since Discovery's course_metadata is the source of truth
for Organization data on edX.org, and pushes updates to
edx-platform's Organization table, it is best that it the
uniqueness is enforced in Discovery's database as well.

TNL-7645