DISCO-1693: Set the Seats on the Course Run when changed
We saw an issue where the official Course Run was not having the
seat price properly updated. It had to do with the course run seats
being used still had the old price because they weren't explicitly
set and the course run was never refreshed from the db. This sets
them and resolves the bug