Ensure editing organizations reactivates them in LMS/Studio
Since LMS/Studio organizations have been completely backfilled,
there are now hundreds of Inactive organizations in the
edxapp database. Inactive organizations are not surfaced to
users, and hence are effectively non-existent.

If an organization is created or edited in Discovery, one
would expect that it is a "real" organization and should
therefore appear as Active in LMS/Studio. This commit
ensures that saving an organization in Discovery tells
LMS to mark the organization as Active if it currently
exists as Inactive.

TNL-7913