feat!: Remove decentralized devstack capability.
See https://open-edx-proposals.readthedocs.io/en/latest/oep-0005-arch-containerize-devstack.html#related-decisions
for more details.  After testing the initial spike we've decided to not
pursue the distributed devstack work.  Our initial spike showed that the
reduction in capabality that resulted from moving to this approach was
too large for the gains it provided.