fix: simplify loading of program data fixtures
In attempting to correct an issue with how newly added translations
were pulled in from the fixture, I noticed that there were a number of
models which are immediately comparable to SeatType and ProgramTypes
which weren't running into whatever issue this code was in place to
solve. I tried removing them to see whether it fixed the loading
issues I was seeing without breaking idempotence, and it did.

JIRA:EDUCATOR-5112