feat!: remove uniqueness on program title field
The removal of unique check on Program title field will mean that title can not be used solely
for fetching the objects from the database. Either marketing_slug or the combination of title and
marketing_slug should be used to fetch the correct Program. Tt is not clear why this check was introduced in the first place.
One potential justification could be that the field "marketing_slug" is auto-generated from the Title field. But marketing_slug is manually entered and has its own uniqueness check in place. But it is not generated from itle, so it is just a hindrance to creating different programs with the same title.