feat!: remove uniqueness on program title field (#3465)
* feat!: remove uniqueness on program title field

The removal of the unique check on the Program title field will mean that the title solely can not be used
for fetching objects from the database. Either marketing_slug or the combination of title and
marketing_slug should be used to fetch the correct Program. 
It is not clear why this check was introduced in the first place. One potential justification could be that the field "marketing_slug" is auto-generated from the Title field. But marketing_slug is manually entered and has its own uniqueness check in place. Since the marketing slug is not generated from the title, the title uniqueness is just a hindrance to creating different programs with the same title.