feat: Update to event-bus-kafka 3.7.1; use its management command directly
This removes the implementation of the `consume_events` management command
that we copied into course-discovery from event-bus-kafka, and now instead
imports the Command from the library.

As of event-bus-kafka 3.7.1, the functionality of the commands is
identical, with one difference: `EVENT_BUS_KAFKA_CONSUMERS_ENABLED` now
defaults to True.

Relevant changes in the library upgrade:

- A fix for the DB reconnection issue.
- Prepares for a breaking change that will remove the `offset_timestamp`
  parameter from `consume_indefinitely`. This is no longer as important
  since that is only called from the management command, but if this
  deletion is rolled back, the management command will need to be re-copied
  from event-bus-kafka to start using the new replay method
  `reset_offsets_and_sleep_indefinitely`.