feat!: Update CSV loader to cater to product source. (#3791)
This update alters the behavior of CSV loader in two aspects. One, the loader now demands that a valid product source should exist so that it can be associated with courses to be ingested. Second, the data presence checks in CSV loader now use a combination of course type and product source to identify which data fields should be provided for successful data ingestion.