feat: add timestamp filtering capabilitiy in Course & Program APIs. (#3771)
FieldTracker is a model utility offered in django-model-utils package keeping track of all changes prior to database commit. For timestamp queries, we are using FieldTracker as a reference to identify if the values in the model are changing or not. Based on changed value, the data modified timestamp in Course or Program will be updated. The data modified timestamp will be used as reference for timestamp filtering.